pyproject.toml
[project]
name = "gemini-cli-agent-team"
version = "0.1.0"
description = "Multi-agent collaborative coding environment powered by Gemini API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.5.0",
    "typer==0.9.0",
    "rich==13.7.0",
    "python-dotenv==1.0.1",
    "pydantic==2.6.4",
    "questionary==2.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
